Official solution, using intervals and sliding window.
Iterate across an array of nhoj's cows and grassy patches combined. If you hit a cow, iterate a sliding window between the cow and the next cow, maintaining a window of size interval_size/2 (This follows the rules).

Then, add the maximum value you can achieve to an array, for both one and two cows.

At the end, just take the top N cows.